refactor: share Wayland window type with remnant
Instead of using the Toplevel class let for wayland::window objects the same
type handle its remnant. Some checks need to be added to special case the
remnant.

+++ win/wayland/space.cpp
@@ -166,7 +166,7 @@ space::~space()
     stacking_order->lock();
 
     for (auto const& window : m_windows) {
-        if (auto win = qobject_cast<win::wayland::window*>(window)) {
+        if (auto win = qobject_cast<win::wayland::window*>(window); win && !win->remnant) {
             destroy_window(win);
             remove_all(m_windows, win);
         }
